Ensure latest version of cranelift is getting used (#47)
A lot of outdated information can be found on the internet. In an environment evolving as fast as the Cranelift project, up-to-date documentation is crucial.

I therefore checked if the simplejit demo works with the latest version of Cranelift (it does) and want to help keeping the repo updates as well.